Tweak evaluation algorithm to ensure we check for all green tiles first
Previously, it was possible to have extra yellow tiles if they were
evaluated before a correctly placed tile.
Tweak to take two loops instead, on the first loop we only check
for green tiles. On the second loop, check all remaining tiles to
determine if they should be yellow or grey.

For example:
  Where the final word is AFTER (`A🟩 F🟩 T🟩 E🟩 R🟩`):
    * TREES should evaluate to `T🟨 R🟨 E⬛ E⬛ S🟩` (where previously it was `T🟨 R🟨 E🟨 E⬛ S🟩`)
     * ENTER should evaluate to `E⬛ N⬛ T🟩 E🟩 R🟩` (where previously it was `E🟨 N⬛ T🟩 E🟩 R🟩`